Went over the interface exposed by create.py and did a bit of experimentation directly from the python interpreter by moving the robot around a room and reading data from the sensors to get an idea of how to get specific sensor data.
A couple of things that stood out are that calling the turn method on a Create
object will stop a previously executing call to the go method. Because of this,
if one would like to actively change the path of the robot while in motion,
the way to do it would be to call go and alter the second argument (degrees/sec)
and correct it over time in order to achieve the desired path. I will add some
code for this on the next run.

I started with some code today, but I'd have to admit I got carried away a bit
on the design aspect at first and ended up writing a rather verbos class that
wraps the Create object. This seems too much for this initial exploratory
programs, but may be necessary later on as I incorporate computer vision, speech
recognition and synthesis, among others... still, the code was written and I
left it that way as I ran out of time. Will spend time tomorrow cleaning up
tomorrow.

Found createviewer.py and played a bit with it, there was a tiny bug caused
by a missing self that - after corrected - allowed me to get feedback in
realtime as the create moved around the room. It's too early to tell how
useful it's going to be in the long run.

--l.